+++ client.go
@@ -43,9 +43,9 @@ type Client struct {
 	// Url is the FMC IP or hostname, e.g. https://10.0.0.1:443 (port is optional).
 	Url string
 	// Authentication token is the current authentication token
-	AuthToken string
+	authToken string
 	// Refresh token is the current authentication token
-	RefreshToken string
+	refreshToken string
 	// UserAgent is the HTTP User-Agent string
 	UserAgent string
 	// Usr is the FMC username. Not used for cdFMC.
@@ -79,7 +79,7 @@ type Client struct {
 	// Rate limit requests to FMC
 	RateLimiterBucket *ratelimit.Bucket
 	// Authentication mutex
-	authenticationMutex *sync.Mutex
+	authenticationMutex *sync.RWMutex
 	// writingMutex protects against concurrent DELETE/POST/PUT requests towards the API.
 	writingMutex *sync.Mutex
 }
@@ -111,8 +111,8 @@ func NewClient(url, usr, pwd string, mods ...func(*Client)) (Client, error) {
 		BackoffMinDelay:     DefaultBackoffMinDelay,
 		BackoffMaxDelay:     DefaultBackoffMaxDelay,
 		BackoffDelayFactor:  DefaultBackoffDelayFactor,
-		authenticationMutex: &sync.Mutex{},
 		RateLimiterBucket:   ratelimit.NewBucketWithRate(1.97, 1), // 1.97 req/s ~= 118 req/min (+/- 1% from 120 req/min that FMC allows)
+		authenticationMutex: &sync.RWMutex{},
 		writingMutex:        &sync.Mutex{},
 	}
 
@@ -279,16 +279,17 @@ func (client *Client) NewReq(method, uri string, body io.Reader, mods ...func(*R
 //	req := client.NewReq("GET", "/api/fmc_config/v1/domain/{DOMAIN_UUID}/object/networks", nil)
 //	res, _ := client.Do(req)
 func (client *Client) Do(req Req) (Res, error) {
-	err := client.Authenticate()
+	err := client.Authenticate("")
 	if err != nil {
 		return Res{}, err
 	}
 
-	// add token
+	// Save current auth token. In case of 401, we can check if the token has changed in the meantime
+	authToken := client.AuthToken()
 	if client.IsCDFMC {
 		req.HttpReq.Header.Add("Authorization", "Bearer "+client.Pwd)
 	} else {
-		req.HttpReq.Header.Add("X-auth-access-token", client.AuthToken)
+		req.HttpReq.Header.Add("X-auth-access-token", authToken)
 	}
 	req.HttpReq.Header.Add("Content-Type", "application/json")
 	req.HttpReq.Header.Add("Accept", "application/json")
@@ -328,7 +329,7 @@ func (client *Client) Do(req Req) (Res, error) {
 		}
 		res = gjson.ParseBytes(bodyBytes)
 		if req.LogPayload {
-			log.Printf("[DEBUG] [ReqID: %s] HTTP Response: %s", req.RequestID, res.Raw)
+			log.Printf("[DEBUG] [ReqID: %s] HTTP Response (StatusCode %d): %s", req.RequestID, httpRes.StatusCode, res.Raw)
 		}
 
 		if httpRes.StatusCode >= 200 && httpRes.StatusCode <= 299 {
@@ -342,19 +343,23 @@ func (client *Client) Do(req Req) (Res, error) {
 			} else if httpRes.StatusCode == 429 || (httpRes.StatusCode >= 500 && httpRes.StatusCode <= 599) {
 				log.Printf("[ERROR] [ReqID: %s] HTTP Request failed: StatusCode %v, Retries: %v", req.RequestID, httpRes.StatusCode, attempts)
 				continue
-			} else if httpRes.StatusCode == 401 && !client.IsCDFMC {
-				// There are bugs in FMC, where the sessions are invalidated out of the blue
-				// In case such a situation is detected, new authentication is forced
-				log.Printf("[DEBUG] [ReqID: %s] Invalid session detected. Forcing reauthentication", req.RequestID)
-				// Clear AuthToken (which is invalid anyways). This also ensures that Authenticate does full authentication
-				client.AuthToken = ""
-				// Force reauthentication, client.Authenticate() takes care of mutexes, hence not calling Login() directly
-				err := client.Authenticate()
-				if err != nil {
-					log.Printf("[DEBUG] [ReqID: %s] HTTP Request failed: StatusCode 401: Forced reauthentication failed: %s", req.RequestID, err.Error())
-					return res, fmt.Errorf("HTTP Request failed: StatusCode 401: Forced reauthentication failed: %s", err.Error())
+			} else if httpRes.StatusCode == 401 {
+				// When request fails with 401, this indicates invalid authentication tokens
+				// For cdFMC, we just retry the request as this may be glitch in cdFMC
+				// For FMC, we try to reauthenticate and retry the request
+				if client.IsCDFMC {
+					log.Printf("[DEBUG] [ReqID: %s] Invalid session detected. Retrying...", req.RequestID)
+					continue
 				}
-				req.HttpReq.Header.Set("X-auth-access-token", client.AuthToken)
+				log.Printf("[DEBUG] [ReqID: %s] Invalid session detected. Reauthenticating...", req.RequestID)
+				err := client.Authenticate(authToken)
+				if err != nil {
+					log.Printf("[DEBUG] [ReqID: %s] HTTP Request failed: StatusCode 401: Reauthentication failed with StatusCode (%d): %s", req.RequestID, httpRes.StatusCode, err.Error())
+					return res, fmt.Errorf("HTTP Request failed: StatusCode 401: Reauthentication failed with StatusCode (%d): %s", httpRes.StatusCode, err.Error())
+				}
+				// Save authentication token in case next 401 is received
+				authToken = client.AuthToken()
+				req.HttpReq.Header.Set("X-auth-access-token", authToken)
 				continue
 			} else if desc := res.Get("error.messages.0.description"); desc.Exists() {
 				// FMC may return HTTP response code 400 with advice to retry the operation
@@ -502,8 +507,9 @@ func (client *Client) Put(path, data string, mods ...func(*Req)) (Res, error) {
 	return client.Do(req)
 }
 
-// Login authenticates to the FMC instance.
-func (client *Client) Login() error {
+// login authenticates to the FMC instance.
+// This function is not thread safe, use Authenticate() instead.
+func (client *Client) login() error {
 	for attempts := 0; ; attempts++ {
 		req, _ := client.NewReq("POST", "/api/fmc_platform/v1/auth/generatetoken", strings.NewReader(""), NoLogPayload)
 		req.HttpReq.Header.Add("User-Agent", client.UserAgent)
@@ -528,8 +534,9 @@ func (client *Client) Login() error {
 				continue
 			}
 		}
-		client.AuthToken = httpRes.Header.Get("X-auth-access-token")
-		client.RefreshToken = httpRes.Header.Get("X-auth-refresh-token")
+
+		client.authToken = httpRes.Header.Get("X-auth-access-token")
+		client.refreshToken = httpRes.Header.Get("X-auth-refresh-token")
 		client.LastRefresh = time.Now()
 		client.RefreshCount = 0
 		client.DomainUUID = httpRes.Header.Get("DOMAIN_UUID")
@@ -547,14 +554,13 @@ func (client *Client) Login() error {
 	}
 }
 
-// Refresh refreshes the authentication token.
-// Note that this will be handled automatically by default.
-// Refresh will be checked every request and the token will be refreshed after 25 minutes.
-func (client *Client) Refresh() error {
+// refresh refreshes the authentication token.
+// This function is not thread safe, use Authenticate() instead.
+func (client *Client) refresh() error {
 	for attempts := 0; ; attempts++ {
 		req, _ := client.NewReq("POST", "/api/fmc_platform/v1/auth/refreshtoken", strings.NewReader(""), NoLogPayload)
-		req.HttpReq.Header.Add("X-auth-access-token", client.AuthToken)
-		req.HttpReq.Header.Add("X-auth-refresh-token", client.RefreshToken)
+		req.HttpReq.Header.Add("X-auth-access-token", client.authToken)
+		req.HttpReq.Header.Add("X-auth-refresh-token", client.refreshToken)
 		req.HttpReq.Header.Add("User-Agent", client.UserAgent)
 		client.RateLimiterBucket.Wait(1)
 		httpRes, err := client.HttpClient.Do(req.HttpReq)
@@ -564,45 +570,77 @@ func (client *Client) Refresh() error {
 		bodyBytes, _ := io.ReadAll(httpRes.Body)
 		httpRes.Body.Close()
 		if httpRes.StatusCode != 204 {
-			log.Printf("[ERROR] Authentication failed: StatusCode %v", httpRes.StatusCode)
-			return fmt.Errorf("authentication failed, status code: %v", httpRes.StatusCode)
+			log.Printf("[ERROR] Authentication token refresh failed: StatusCode %v", httpRes.StatusCode)
+			return fmt.Errorf("authentication token refresh failed, status code: %v", httpRes.StatusCode)
 		}
 		if len(bodyBytes) > 0 {
 			if ok := client.Backoff(attempts); !ok {
-				log.Printf("[ERROR] Authentication failed: Invalid credentials")
-				return fmt.Errorf("authentication failed, invalid credentials")
+				log.Printf("[ERROR] Authentication token refresh failed: Invalid credentials")
+				return fmt.Errorf("authentication token refresh failed, invalid credentials")
 			} else {
-				log.Printf("[ERROR] Authentication failed: %s, retries: %v", err, attempts)
+				log.Printf("[ERROR] Authentication token refresh failed: %s, retries: %v", err, attempts)
 				continue
 			}
 		}
-		client.AuthToken = httpRes.Header.Get("X-auth-access-token")
-		client.RefreshToken = httpRes.Header.Get("X-auth-refresh-token")
+
+		client.authToken = httpRes.Header.Get("X-auth-access-token")
+		client.refreshToken = httpRes.Header.Get("X-auth-refresh-token")
 		client.LastRefresh = time.Now()
 		client.RefreshCount = client.RefreshCount + 1
 		client.DomainUUID = httpRes.Header.Get("DOMAIN_UUID")
-		log.Printf("[DEBUG] Refresh successful")
+
+		log.Printf("[DEBUG] Authentication token refresh successful")
 		return nil
 	}
 }
 
-// Login if no token available.
-func (client *Client) Authenticate() error {
-	// cdFMC uses fixed token, no need to do separate authentication
+// AuthToken returns the current token
+func (client *Client) AuthToken() string {
+	client.authenticationMutex.RLock()
+	defer client.authenticationMutex.RUnlock()
+
+	return client.authToken
+}
+
+// Authenticate assures the token is there and valid.
+// It will try to login/refresh the token based on the current state and information
+// from FMC on failures (no proactive reauthentications).
+// currentAuthToken is the token used in the request. This helps to
+// determine, if authToken needs refreshing or has already been refreshed by other thread.
+// currentAuthToken can be an empty string.
+func (client *Client) Authenticate(currentAuthToken string) error {
+	// cdFMC uses fixed token to authenticate
 	if client.IsCDFMC {
 		return nil
 	}
 
-	var err error
 	client.authenticationMutex.Lock()
-	// Check if we can attempt to refresh the token (there is old token, it's between 25 and 29 minutes since last refresh, and less than 3 refreshes done)
-	if client.AuthToken != "" && time.Since(client.LastRefresh) > 1500*time.Second && time.Since(client.LastRefresh) < 1740*time.Second && client.RefreshCount < 3 {
-		err = client.Refresh()
-		// Check if we need to login (no token available or more than 25 minutes since last refresh)
-	} else if client.AuthToken == "" || time.Since(client.LastRefresh) >= 1500*time.Second {
-		err = client.Login()
+	defer client.authenticationMutex.Unlock()
+
+	if client.authToken != "" && currentAuthToken == "" {
+		// authToken is present, no error reported, do nothing
+		return nil
 	}
-	client.authenticationMutex.Unlock()
+
+	if currentAuthToken != "" && currentAuthToken != client.authToken {
+		// authToken has changed since the last request
+		// we assume some other thread has already refreshed it, do nothing
+		return nil
+	}
+
+	if client.authToken == "" {
+		// No authToken, do a full login
+		return client.login()
+	}
+
+	// We have the tokens, but FMC rejected them
+	// first check if we can refresh the tokens
+	err := client.refresh()
+	if err != nil {
+		// If refresh fails, do a full login
+		err = client.login()
+	}
+
 	return err
 }