+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.2 (Unreleased)
 
+- BREAKING CHANGE: `login()` and `refresh()` functions are no logner exported
+- BREAKING CHANGE: `Authenticate()` is now defined as `Authenticate(failedAuthToken string)`
+- Fix: Possible race-conditions and miss-handling of the FMC authentication tokens
 - Enh: Introduced ReqID to correlate events in the logs
 - Enh: FMC 7.4.1, 7.6.0 and later releases have rate-limit increased to 300 req/min